pyproject.toml
[project]
name = "test-ollama-cloud"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "ollama>=0.6.0",
    "python-dotenv>=1.1.1",
]
